fix: resolve circular dependency in /lib/api/helpers by moving getShopLang and getSlug to client/modules/core/main.js
