+++ lib/api/helpers.js
@@ -18,18 +18,6 @@ if (Meteor.isClient) {
  */
 
 /**
- * @name getShopLang
- * @method
- * @memberof Helpers
- * @return {String} returns current shop language
- */
-export function getShopLang() {
-  const shopId = Reaction.getShopId();
-  const shop = Shops.findOne({ _id: shopId });
-  return shop ? shop.language : "";
-}
-
-/**
  * @name toCamelCase
  * @method
  * @memberof Helpers
